+++ R/bill.R
@@ -52,7 +52,7 @@ cong_bill <- function(congress = NULL, type = NULL, number = NULL, item = NULL,
     httr2::req_headers(
       "accept" = glue::glue("application/{format}")
     )
-  out <- req |>
+  resp <- req |>
     httr2::req_perform()
 
   formatter <- switch(format,
@@ -60,7 +60,7 @@ cong_bill <- function(congress = NULL, type = NULL, number = NULL, item = NULL,
                       'xml' = httr2::resp_body_xml
   )
 
-  out <- out |>
+  out <- resp <- resp |>
     formatter()
 
   if (clean) {
@@ -75,16 +75,39 @@ cong_bill <- function(congress = NULL, type = NULL, number = NULL, item = NULL,
           purrr::pluck('bill') |>
           tibble::enframe() |>
           tidyr::pivot_wider() |>
-          tidyr::unnest_wider(col = where(~purrr::vec_depth(.x) < 4), simplify = TRUE, names_sep = '_') |>
+          tidyr::unnest_wider(col = where(~purrr::pluck_depth(.x) < 4), simplify = TRUE, names_sep = '_') |>
           dplyr::rename_with(.fn = function(x) stringr::str_sub(x, end = -3), .cols = dplyr::ends_with('_1')) |>
           clean_names()
       } else {
-        out <- out |>
-          purrr::pluck(item) |>
-          dplyr::bind_rows() |>
-          clean_names()
+        if (item == 'text')  {
+          item <- 'textVersions'
+          out <- out |>
+            purrr::pluck(item)
+
+          out <- lapply(out, function(x) {
+            x[[2]] <- x$formats |>
+              dplyr::bind_rows() |>
+              tidyr::pivot_wider(names_from = 'type', values_from = 'url') |>
+              dplyr::rename_with(.fn = function(x) paste0('url_', str_colname(x)))
+            x
+          })
+
+          out <- out |>
+            dplyr::bind_rows() |>
+            tidyr::unnest_wider(col = 'formats') |>
+            clean_names()
+        } else {
+          out <- out |>
+            purrr::pluck(item) |>
+            list_hoist() |>
+            clean_names()
+       }
+
       }
     }
+    out <- out |>
+      add_resp_info(resp) |>
+      cast_date_columns()
   }
   out
 }
@@ -107,5 +130,5 @@ bill_endpoint <- function(congress, type, number, item) {
     }
   }
 
-  out
+  tolower(out)
 }