+++ R/congress.R
@@ -47,7 +47,7 @@ cong_congress <- function(congress = NULL,
       "accept" = glue::glue("application/{format}")
     )
 
-  out <- req |>
+  resp <- req |>
     httr2::req_perform()
 
   formatter <- switch(format,
@@ -55,7 +55,7 @@ cong_congress <- function(congress = NULL,
                       'xml' = httr2::resp_body_xml
   )
 
-  out <- out |>
+  out <- resp <- resp |>
     formatter()
 
   if (clean) {
@@ -70,10 +70,13 @@ cong_congress <- function(congress = NULL,
         purrr::pluck('congress') |>
         tibble::enframe() |>
         tidyr::pivot_wider() |>
-        tidyr::unnest_wider(col = where(~purrr::vec_depth(.x) < 4), simplify = TRUE, names_sep = '_') |>
+        tidyr::unnest_wider(col = where(~purrr::pluck_depth(.x) < 4), simplify = TRUE, names_sep = '_') |>
         dplyr::rename_with(.fn = function(x) stringr::str_sub(x, end = -3), .cols = dplyr::ends_with('_1')) |>
         clean_names()
     }
+    out <- out |>
+      add_resp_info(resp) |>
+      cast_date_columns()
   }
   out
 }
@@ -84,5 +87,5 @@ congress_endpoint <- function(congress, type) {
     out <- paste0(out, '/', congress)
   }
 
-  out
+  tolower(out)
 }