+++ R/nomination.R
@@ -18,7 +18,7 @@
 #'
 #' cong_nomination()
 #'
-#' cong_nomination(congress = 117)
+#' cong_nomination(congress = 118)
 #'
 #' cong_nomination(congress = 117, number = 2467)
 #'
@@ -53,7 +53,7 @@ cong_nomination <- function(congress = NULL, number = NULL, item = NULL,
     httr2::req_headers(
       "accept" = glue::glue("application/{format}")
     )
-  out <- req |>
+  resp <- req |>
     httr2::req_perform()
 
   formatter <- switch(format,
@@ -61,7 +61,7 @@ cong_nomination <- function(congress = NULL, number = NULL, item = NULL,
                       'xml' = httr2::resp_body_xml
   )
 
-  out <- out |>
+  out <- resp <- resp |>
     formatter()
 
   if (clean) {
@@ -76,7 +76,7 @@ cong_nomination <- function(congress = NULL, number = NULL, item = NULL,
           purrr::pluck('nomination') |>
           tibble::enframe() |>
           tidyr::pivot_wider() |>
-          tidyr::unnest_wider(col = where(~purrr::vec_depth(.x) < 4), simplify = TRUE, names_sep = '_') |>
+          tidyr::unnest_wider(col = where(~purrr::pluck_depth(.x) < 4), simplify = TRUE, names_sep = '_') |>
           dplyr::rename_with(.fn = function(x) stringr::str_sub(x, end = -3), .cols = dplyr::ends_with('_1')) |>
           clean_names()
       } else {
@@ -86,6 +86,9 @@ cong_nomination <- function(congress = NULL, number = NULL, item = NULL,
           clean_names()
       }
     }
+    out <- out |>
+      add_resp_info(resp) |>
+      cast_date_columns()
   }
   out
 }
@@ -104,5 +107,5 @@ nomination_endpoint <- function(congress, number, item) {
     }
   }
 
-  out
+  tolower(out)
 }