+++ R/record.R
@@ -1,4 +1,4 @@
-#' Request Bill Information
+#' Request Congressional Record Information
 #'
 #' @param year integer for year
 #' @param month integer for month
@@ -40,7 +40,7 @@ cong_record <- function(year = NULL, month = NULL, day = NULL,
     httr2::req_headers(
       "accept" = glue::glue("application/{format}")
     )
-  out <- req |>
+  resp <- req |>
     httr2::req_perform()
 
   formatter <- switch(format,
@@ -48,7 +48,7 @@ cong_record <- function(year = NULL, month = NULL, day = NULL,
                       'xml' = httr2::resp_body_xml
   )
 
-  out <- out |>
+  out <- resp <- resp |>
     formatter()
 
   if (clean) {
@@ -56,9 +56,13 @@ cong_record <- function(year = NULL, month = NULL, day = NULL,
       purrr::pluck('Results') |>
       tibble::enframe() |>
       tidyr::pivot_wider() |>
-      tidyr::unnest_wider(col = where(~purrr::vec_depth(.x) < 4), simplify = TRUE, names_sep = '_') |>
+      tidyr::unnest_wider(col = where(~purrr::pluck_depth(.x) < 4), simplify = TRUE, names_sep = '_') |>
       dplyr::rename_with(.fn = function(x) stringr::str_sub(x, end = -3), .cols = dplyr::ends_with('_1')) |>
       clean_names()
+
+    out <- out |>
+      add_resp_info(resp) |>
+      cast_date_columns()
   }
   out
 }
@@ -67,5 +71,5 @@ cong_record <- function(year = NULL, month = NULL, day = NULL,
 record_endpoint <- function() {
   out <- 'congressional-record'
 
-  out
+  tolower(out)
 }